+++ Forms/CommandLineForm.vb
@@ -199,7 +199,7 @@ Public Class CommandLineForm
 
                 Dim max = oParam.Options.Select(Function(txt) txt.Length).Max
 
-                If max > 23 Then
+                If max > 24 Then
                     menuBlock.Button.Expand = True
                 End If